style(view): Application.java에 있던 뷰를 뷰 패키지 만들어서 분리
변경 전
- Application 내부에 Calculate View 존재

변경 후
- calculator 패키지에 view 패키지를 만들고, 그 안에 CalculateView.java 클래스를 만들어 input 함수와 output 함수 구현
