+++ src/main/java/calculator/Application.java
@@ -2,19 +2,7 @@ package calculator;
 
 import calculator.model.Calculator;
 import calculator.model.Expression;
-import camp.nextstep.edu.missionutils.Console;
-
-// view
-class CalculateView {
-    public String getUserInput() {
-        System.out.println("덧셈할 문자열을 입력해 주세요.");
-        return Console.readLine();
-    }
-
-    public void displayResult(int cnt) {
-        System.out.println("결과 : " + cnt);
-    }
-}
+import calculator.view.CalculateView;
 
 // controller
 class CalculatorController {
