+++ src/main/java/calculator/view/CalculateView.java
@@ -1,0 +1,14 @@
+package calculator.view;
+
+import camp.nextstep.edu.missionutils.Console;
+
+public class CalculateView {
+    public String getUserInput() {
+        System.out.println("덧셈할 문자열을 입력해 주세요.");
+        return Console.readLine();
+    }
+
+    public void displayResult(int cnt) {
+        System.out.println("결과 : " + cnt);
+    }
+}
